Extract GameBoardExtensions to own file, add tests.
